extract event visitors to dfferent file
